requirements.txt
langchain==0.3.26
pypdf==5.6.1
sentence-transformers==4.1.0
flask==3.1.1
langchain-pinecone==0.2.8
python-dotenv==1.1.0
langchain-community==0.3.26
langchain-google-genai
google-generativeai
-e .